+++ CHANGELOG.md
@@ -6,6 +6,8 @@
    or goes into them. It will prevent a potential freezing bug.
  - `Response::Waker` associated type is removed.
  - `std-futures` feature is renamed to `std-future`.
+ - Can convert any type of `Response` into futures 0.1 `Future`, or
+   `std::future::Future`.
 
 
 0.1.0 (March 22, 2019)